# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/question/": {
            "get": {
                "description": "Fetches the next available question for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Get a Question",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.GetQuestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/question.FullError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/question.FullError"
                        }
                    }
                }
            },
            "post": {
                "description": "Allows a user to submit an answer to a question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Submit an Answer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "User answer submission",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/question.AnswerQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.AnswerQuestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/question.FullError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/question.FullError"
                        }
                    }
                }
            }
        },
        "/rank/": {
            "get": {
                "description": "Retrieves the overall leaderboard",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rank"
                ],
                "summary": "Get Leaderboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rank.GetLeaderboardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rank.FullError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rank.FullError"
                        }
                    }
                }
            }
        },
        "/rank/player/": {
            "get": {
                "description": "Fetches the ranking position of a player",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rank"
                ],
                "summary": "Get Player Position",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rank.GetPlayerPositionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rank.FullError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rank.FullError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "question.AnswerQuestionRequest": {
            "type": "object",
            "required": [
                "answer",
                "number_question"
            ],
            "properties": {
                "answer": {
                    "type": "integer"
                },
                "number_question": {
                    "type": "integer"
                }
            }
        },
        "question.AnswerQuestionResponse": {
            "type": "object",
            "properties": {
                "is_correct": {
                    "type": "boolean"
                }
            }
        },
        "question.FullError": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "question.GetQuestionResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "number_question": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "rank.FullError": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "rank.GetLeaderboardResponse": {
            "type": "object",
            "properties": {
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rank.PlayerPosition"
                    }
                }
            }
        },
        "rank.GetPlayerPositionResponse": {
            "type": "object",
            "properties": {
                "player": {
                    "$ref": "#/definitions/rank.PlayerPosition"
                }
            }
        },
        "rank.PlayerPosition": {
            "type": "object",
            "properties": {
                "position": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}